+++ src/models/usersDao.js
@@ -53,7 +53,37 @@ const getUserByEmail = async (email) => {
   }
 };
 
+const addUserAddress = async (userId, address, postalCode) => {
+  try {
+    const [user] = await dataSource.query(
+      `SELECT id, address, postalcode AS postalCode
+      FROM users
+      WHERE id = ?`,
+      [userId]
+    );
+
+    if (user) {
+      const user_id = user.id;
+
+      await dataSource.query(
+        `UPDATE users
+        SET
+          address = ?,
+          postalcode = ?
+        WHERE id = ?`,
+        [address, postalCode, user_id]
+      );
+    } else {
+      throw new Error("User not found");
+    }
+  } catch (err) {
+    console.log(err);
+    throw new Error("Error updating address for User usersDAO " + err.message);
+  }
+};
+
 module.exports = {
   signUp,
   getUserByEmail,
+  addUserAddress
 };
