+++ src/routes/ordersRoute.js
@@ -4,7 +4,6 @@ const { checkToken } = require("../middlewares/auth")
 
 const router = express.Router();
 
-router.post("", checkToken, ordersController.addUserAddress);
 
 module.exports = {
   router,
